v1.16.17 (ca8eecab367b76556380378afcbd8a9459e69015713f642258028581f0d721b74770b759be77fe8eed19da68d293848532504daa185729269dff2360ed4467ac)

+++ Qwilight/System/AudioSystem/AudioSystem.cs
@@ -164,7 +164,7 @@ namespace Qwilight
 
         public ConcurrentDictionary<string, AudioItem> DefaultAudioItemMap { get; } = new();
 
-        public string GetDefaultAudioFileName(int randomMillis)
+        public string GetDefaultAudioFileName(long randomMillis)
         {
             var defaultAudioFilePathItems = Configure.Instance.DefaultAudioFilePathItems;
             var defaultAudioFilePathItemsLength = defaultAudioFilePathItems.Length;
